backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
